% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{package}
\name{AtlasMaker-package}
\alias{AtlasMaker}
\alias{AtlasMaker-package}
\title{AtlasMaker: Easily Make Multiple Leaflet Maps in a Shiny Application}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rachel-greenlee/AtlasMaker}
  \item Report bugs at \url{https://github.com/rachel-greenlee/AtlasMaker/issues}
}

}
\author{
\strong{Maintainer}: Rachel Greenlee \email{first.last@example.com}

Authors:
\itemize{
  \item Zachary Palmore \email{first.last@example.com}
}

Other contributors:
\itemize{
  \item Jason Bryer \email{jason.bryer@cuny.edu} [contributor]
  \item Angela Lui \email{angela.lui@cuny.edu} [contributor]
}

}
\keyword{internal}
