% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{birds}
\alias{birds}
\title{birds}
\format{
A spatial data frame with 62 county entries:
\describe{
\item{STATEP}{state's code}
\item{NAME}{county name}
\item{INTPLAT}{latitude}
\item{INTPTLON}{longitude}
\item{ALAND}{area of land, in square meters}
\item{AWATER}{area of water, in square meters}
\item{Taxonomic.Group}{For animals and plants, the taxonomic phylum, class, or order to which the species belongs.}
\item{fill_value}{count of species from birds taxonomic group by county}
...
}
}
\source{
\url{https://data.ny.gov/Energy-Environment/Biodiversity-by-County-Distribution-of-Animals-Pla/tk82-7km5}
}
\description{
Derived from data.ny.gov biodiversity data by county, filtered for birds only.
}
