% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{counties_NY}
\alias{counties_NY}
\title{counties_NY}
\format{
A data frame with X rows and Y variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{}
}
\description{
Say more
}
