% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker.R
\name{map_server}
\alias{map_server}
\title{The Shiny server module for AtlasMaker.}
\usage{
map_server(
  id,
  polygons = NULL,
  points = NULL,
  polylines = NULL,
  pal = NULL,
  center = NULL,
  min_zoom = 7,
  map_base_theme = "Stamen.Terrain",
  poly_palette = "BuPu",
  point_color = "black",
  polyline_color = "gray"
)
}
\arguments{
\item{unique}{identifier for the map. This must match the id used in \code{\link[=map_UI]{map_UI()}}.}
}
\description{
DESCRIBE HERE...
}
