% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_server.R
\name{map_UI}
\alias{map_UI}
\title{The user interface for AtlasMaker.}
\usage{
map_UI(id)
}
\value{
a \code{\link[leaflet:map-shiny]{leaflet::leafletOutput()}} object.
}
\description{
This is the core function for AtlasMaker where users pass in the spatial-data
based lists and aesthetic choices for each map tab.
}
