% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{points}
\alias{points}
\title{points}
\format{
A function with arguments x, ...:
\describe{
\item{x}{user data}
...
}
}
\source{
\url{none}
}
\description{
Takes user defined data for points in demo1
Default is NULL or empty
}
