% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{points_parks}
\alias{points_parks}
\title{points_parks}
\format{
A data frame with 254 rows and 3 variables:
\describe{
\item{label}{name of state park}
\item{long}{longitude value}
\item{lat}{latitude value}
...
}
}
\source{
\url{https://data.ny.gov/Recreation/Watchable-Wildlife-Sites/hg7a-5ssi}
}
\description{
Data on state parks in New  York. For this package demo only three columns were retained.
}
