% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{points_watchsites}
\alias{points_watchsites}
\title{points_watchsites}
\format{
A data frame with 76 rows and 3 variables:
\describe{
\item{label}{name of watchsite location}
\item{long}{longitude value}
\item{lat}{latitude value}
...
}
}
\source{
\url{https://data.ny.gov/Recreation/Watchable-Wildlife-Sites/hg7a-5ssi}
}
\description{
Data originally from data.ny.gov, for this package demo only three columns were retained.
}
