% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{roads_ny_interstate}
\alias{roads_ny_interstate}
\title{roads_ny_interstate}
\format{
A data frame with 245 rows and 5 variables:
\describe{
\item{LINEARID}{unique identifier}
\item{FULLNAME}{interstate name}
\item{RTTYP}{route type}
\item{MTFCC}{US Census feature class code}
\item{geometry}{polyline information}
...
}
}
\description{
US Census polyline data for New York state interstates accessed through the \code{tigris} R Package.
}
