% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_AtlasMaker.R
\name{shiny_AtlasMaker}
\alias{shiny_AtlasMaker}
\title{Run a the AtlasMaker Shiny Demo}
\usage{
shiny_AtlasMaker(app = c("demo1"))
}
\arguments{
\item{app}{defaults to demo1}
}
\value{
Demo of AtlasMaker, a Shiny app that displays 4 tabs of Leaflet maps. See package vignette for code.
}
\description{
Run a the AtlasMaker Shiny Demo
}
