% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{package}
\name{AtlasMaker-package}
\alias{AtlasMaker}
\alias{AtlasMaker-package}
\title{AtlasMaker: Make Multiple 'leaflet' Maps in 'Shiny'}
\description{
Simplify creating multiple, related 'leaflet' maps across tabs for a 'shiny' application. Users build lists of any polygons, points, and polylines needed for the project, use the map_server() function to assign built lists and other chosen aesthetics into each tab, and the package leverages modules to generate all map tabs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rachel-greenlee/AtlasMaker}
  \item Report bugs at \url{https://github.com/rachel-greenlee/AtlasMaker/issues}
}

}
\author{
\strong{Maintainer}: Rachel Greenlee \email{rachellynn.greenlee@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Zachary Palmore \email{zacharypalmore@gmail.com}
}

Other contributors:
\itemize{
  \item Jason Bryer \email{jason.bryer@cuny.edu} [contributor]
  \item Angela Lui \email{angela.lui@cuny.edu} [contributor]
}

}
\keyword{internal}
