% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{package}
\name{AtlasMaker-package}
\alias{AtlasMaker}
\alias{AtlasMaker-package}
\title{AtlasMaker: Easily Make Multiple Leaflet Maps in a Shiny Application}
\description{
The goal of AtlasMaker is to create multiple, related `leaflet` maps across tabs for a `shiny` application without having to copy, paste, and continually edit core code blocks that share much of the same structure across map tabs. Users build lists of any polygons, points, and polylines needed for the project, use the `map_server()` function to feed built lists and other chosen aesthetics into the tab, and the package does the rest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rachel-greenlee/AtlasMaker}
  \item Report bugs at \url{https://github.com/rachel-greenlee/AtlasMaker/issues}
}

}
\author{
\strong{Maintainer}: Rachel Greenlee \email{rachellynn.greenlee@gmail.com}

Authors:
\itemize{
  \item Zachary Palmore \email{first.last@example.com}
}

Other contributors:
\itemize{
  \item Jason Bryer \email{jason.bryer@cuny.edu} [contributor]
  \item Angela Lui \email{angela.lui@cuny.edu} [contributor]
}

}
\keyword{internal}
