% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{counties_NY}
\alias{counties_NY}
\title{counties_NY}
\format{
A spatial data frame with 62 rows and 7 variables:
\describe{
\item{STATEP}{state's code}
\item{NAME}{county name}
\item{INTPLAT}{latitude}
\item{INTPTLON}{longitude}
\item{ALAND}{area of land, in square meters}
\item{AWATER}{area of water, in square meters}
\item{geometry}{polygon information}
...
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\description{
US Census polygon data for New York state counties accessed through the \code{tigris} R Package.
}
