% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker.R
\name{map_UI}
\alias{map_UI}
\title{The user interface for AtlasMaker.}
\usage{
map_UI(id)
}
\arguments{
\item{id}{unique identifier for the map. This must match the id used in \code{\link[=map_server]{map_server()}}.}
}
\value{
a \code{\link[leaflet:map-shiny]{leaflet::leafletOutput()}} object.
}
\description{
Describe here...
}
