% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasMaker-package.R
\docType{data}
\name{points_campgrounds}
\alias{points_campgrounds}
\title{points_campgrounds}
\format{
A data frame with 116 rows and 3 variables:
\describe{
\item{label}{name of the campground}
\item{long}{longitude value}
\item{lat}{latitude value}
...
}
}
\source{
\url{https://data.ny.gov/Recreation/Campgrounds-by-County-Within-Adirondack-Catskill-F/tnqf-vydw}

\url{https://data.ny.gov/Recreation/Campgrounds-by-County-Outside-Adirondack-Catskill-/5zxz-z3ci}
}
\description{
Data on campgrounds in and outside of the Adirondacks in New York.
For this package demo two data files were combined and only three columns were retained.
}
